Update file discard confirm key to uppercase 'D'
This is done for issue #2291. It is unwise to have a discard confirm keybind
to be the same as the one that inits it. It can happen by mistake to have a double
click which could result in a big result of work. This way we have to put a bit more
effort into achieving it, making us more careful.